refactor: deduplicate entity seeding
